fix(OpenGL): removing unused components & adding missing create system
