+++ src/plugin/opengl/src/OpenGL.hpp
@@ -3,7 +3,6 @@
 #include "component/Model/Model.hpp"
 #include "plugin/PluginOpenGL.hpp"
 #include "resource/Camera/Camera.hpp"
-#include "resource/GLFWWindow/GLFWWindow.hpp"
 #include "resource/MaterialCache/MaterialCache.hpp"
 #include "resource/ShaderManager/ShaderManager.hpp"
 #include "system/AllSystems.hpp"
