+++ src/plugin/opengl/xmake.lua
@@ -29,7 +29,6 @@ target("PluginOpenGL")
     add_includedirs("src/utils/Light", {public = true})
     add_includedirs("src/component/Model/", {public = true})
     add_includedirs("src/resource/ShaderManager/", {public = true})
-    add_includedirs("src/resource/GLFWWindow/", {public = true})
     add_includedirs("src/resource/MaterialCache/", {public = true})
     add_includedirs("src/resource/Camera/", {public = true})
     add_includedirs("src/system/", {public = true})
