--- src/plugin/window/src/plugin/PluginWindow.hpp
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "APlugin.hpp"
-
-namespace ES::Plugin::Window {
-class Plugin : public ES::Engine::APlugin {
-  public:
-    explicit Plugin(ES::Engine::Core &core)
-        : ES::Engine::APlugin(core){
-              // empty
-          };
-    ~Plugin() = default;
-
-    void Bind() final;
-};
-} // namespace ES::Plugin::Window
